feat:canviem component per generar documentació amb swagger
