--- api-v1/errors.js
@@ -1,16 +1,0 @@
-{
-  [
-    {
-      // A URI reference [RFC3986] that identifies the problem type.
-      type: '',
-      // A short, human-readable summary of the problem type.
-      title: '',
-      // The HTTP status code
-      status: '',
-      // A human-readable explanation specific to this occurrence of the problem
-      detail: '',
-      // A URI reference that identifies the specific occurrence of the problem
-      instance: '',
-    },
-  ]
-}
