+++ doc/adr/2023-03-06 openapi.md
@@ -1,0 +1,53 @@
+# Sistema de documentació d'APIs
+
+Per tal de poder desenvolupar i documentar APIs de forma fàcil i evitant que estigui desactualitzada davant de canvis, podem fer servir diversos sistemes i hem de decidir quin s'adpta millor a la nostra filosofia.
+
+- Staus: [proposed | in progress | rejected | accepted | deprecated by | obsoletes | continued by | extended by [ADR-0005-Example](0005-example.md)
+- Date: [YYYY-MM-DD when the status changes occurred]
+- Authors:
+- Deciders:
+- Related documments:
+  - [f.i. Wiki documents, architecture diagrams, external links, a pdf,
+     if needed use Bibliographical quote style like http://www.citethisforme.com/citation-generator/harvard]
+
+## Context and Problem Statement
+
+Desenvolupem APIs REST amb serveis amb NodeJS.
+Es vol documentar les APIs de forma coherent i evitant que quedin desactualitzades davant de canvis.
+
+## Decision
+
+### Decision Drivers
+
+- Simplicitat d'ús
+- Facilitat de manteniment: Reducció de dependpències externes
+- Flexibilitat a l'hora de codificar
+- Reducció de les possibilitats de desactualitzar la documentació
+- Qualitat de la documentació
+
+### Considered Options
+
+1. Apidoc.js (<https://apidocjs.com/>): permet crear documentació a partir d'anotacions en el codi font, no fa cap validació dels objectes i crides segons la documentació.
+2. Swagger amb express-openapi: genera la documentació fent servir swagger (<https://swagger.io/>) i valida en les crides que les peticions s'adaptin als objectes definits.
+3. Swagger amb swagger-jsdoc: genera la documentació fent servir swagger (<https://swagger.io/>) i valida en les crides que les peticions s'adaptin als objectes definits.
+
+### Decision Outcome
+
+Triem l'opció 3 ja que compleix la majoria de criteris, la 1a es queda curta a nivell de documentació ja que no permet fer tests, la 2a és massa rígida i la 3a genera bona documentació sense limitar l'estrucutra del codi.
+
+#### Pros and cons and risks expected benefits
+
+| Decision drivers | Opció 1 | Opció 2 | Opció 3 | Comentaris |
+|-|-|-|-|-|
+| Simplicitat | ++ | -- | + | La 1a és més simple que la 2a i 3a (només cal posar anotacions i re/genera la documentació, no és gens invasiva, les altres ens obliguen a afegir dependències al codi) |
+| Mantenibilitat | + | - | + | La 1a només té una dependència i està mantinguda per la comunitat, no ens obliga a canviar l'estructura del codi, deixar de fer-la servir no ens obliga a canviar el codi, les altres tenen 2 dependpències. La 2a i 3a fa servir swagger que és un estandard actualment |
+| Flexibilitat | + | - | + | La 1a i 3a no ens limita la manera de codificar ni ens imposa cap estructura, en canvi la 2a, ens obliga a crear documents en carpetes concretes, ens valida de forma oculta les crides segons les api definides |
+| Evitar desactualitzar doc | - | + | + | La 2a i 3a opció ens valida millor la coherència de la documentació respecte al codi. En la 1a s'ha d'actualitzar l'app generaadora de documentació cada vegada que hi ha un canvi |
+| Qualitat doc. | - | + | + | La 2a i 3a opció genera documentació de més qualitat amb possibilitat d'executar test |
+
+### Links
+
+- <https://apidocjs.com/>
+- <https://www.npmjs.com/package/swagger-ui-express>
+- <https://medium.com/@italo.ravina/a%C3%B1adir-documentaci%C3%B3n-con-swagger-a-un-api-creada-en-express-5c4c5c3cb19e>
+- <https://adr.github.io/>
