+++ package.json
@@ -8,14 +8,15 @@
     "nodemon": "nodemon ./bin/www"
   },
   "dependencies": {
+    "body-parser": "^1.20.2",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
-    "express": "~4.16.1",
-    "express-openapi": "^12.0.2",
+    "express": "^4.18.2",
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1",
     "pug": "^3.0.2",
-    "swagger-ui-express": "^4.6.0",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^4.6.2",
     "winston": "^3.8.2"
   },
   "devDependencies": {
